+++ internal/storage/v1/elasticsearch/spanstore/reader.go
@@ -10,6 +10,7 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
+	"strconv"
 	"time"
 
 	"github.com/olivere/elastic"
@@ -19,6 +20,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger-idl/model/v1"
+	"github.com/jaegertracing/jaeger/internal/storage/v1/api/spanstore"
 	"github.com/jaegertracing/jaeger/internal/storage/v1/elasticsearch/spanstore/internal/dbmodel"
 	"github.com/jaegertracing/jaeger/pkg/es"
 	cfg "github.com/jaegertracing/jaeger/pkg/es/config"
@@ -74,6 +76,22 @@ var (
 
 	nestedTagFieldList = []string{nestedTagsField, nestedProcessTagsField, nestedLogFieldsField}
 )
+
+// SpanReaderV1	is a wrapper around SpanReader
+type SpanReaderV1 struct {
+	spanReader    *SpanReader
+	spanConverter dbmodel.ToDomain
+	tracer        trace.Tracer
+}
+
+// NewSpanReaderV1 returns an instance of SpanReaderV1
+func NewSpanReaderV1(p SpanReaderParams) *SpanReaderV1 {
+	return &SpanReaderV1{
+		spanReader:    NewSpanReader(p),
+		spanConverter: dbmodel.NewToDomain(p.TagDotReplacement),
+		tracer:        p.Tracer,
+	}
+}
 
 // SpanReader can query for and load traces from ElasticSearch
 type SpanReader struct {
@@ -220,6 +238,22 @@ func timeRangeIndices(indexName, indexDateLayout string, startTime time.Time, en
 	return indices
 }
 
+// GetTrace takes a traceID and returns a Trace associated with that traceID
+func (s *SpanReaderV1) GetTrace(ctx context.Context, query spanstore.GetTraceParameters) (*model.Trace, error) {
+	ctx, span := s.spanReader.tracer.Start(ctx, "GetTrace")
+	defer span.End()
+	currentTime := time.Now()
+	// TODO: use start time & end time in "query" struct
+	traces, err := s.multiRead(ctx, []model.TraceID{query.TraceID}, currentTime.Add(-s.spanReader.maxSpanAge), currentTime)
+	if err != nil {
+		return nil, es.DetailedError(err)
+	}
+	if len(traces) == 0 {
+		return nil, spanstore.ErrTraceNotFound
+	}
+	return traces[0], nil
+}
+
 func (s *SpanReader) collectJsonSpans(esSpansRaw []*elastic.SearchHit) ([]*dbmodel.Span, error) {
 	spans := make([]*dbmodel.Span, len(esSpansRaw))
 	for i, esSpanRaw := range esSpansRaw {
@@ -228,6 +262,19 @@ func (s *SpanReader) collectJsonSpans(esSpansRaw []*elastic.SearchHit) ([]*dbmod
 			return nil, fmt.Errorf("marshalling JSON to span object failed: %w", err)
 		}
 		spans[i] = jsonSpan
+	}
+	return spans, nil
+}
+
+func (s *SpanReaderV1) collectSpans(jsonSpans []*dbmodel.Span) ([]*model.Span, error) {
+	spans := make([]*model.Span, len(jsonSpans))
+
+	for i, jsonSpan := range jsonSpans {
+		span, err := s.spanConverter.SpanToDomain(jsonSpan)
+		if err != nil {
+			return nil, fmt.Errorf("converting JSONSpan to domain Span failed: %w", err)
+		}
+		spans[i] = span
 	}
 	return spans, nil
 }
@@ -246,6 +293,11 @@ func (*SpanReader) unmarshalJSONSpan(esSpanRaw *elastic.SearchHit) (*dbmodel.Spa
 }
 
 // GetServices returns all services traced by Jaeger, ordered by frequency
+func (s *SpanReaderV1) GetServices(ctx context.Context) ([]string, error) {
+	return s.spanReader.GetServices(ctx)
+}
+
+// GetServices returns all services traced by Jaeger, ordered by frequency
 func (s *SpanReader) GetServices(ctx context.Context) ([]string, error) {
 	ctx, span := s.tracer.Start(ctx, "GetService")
 	defer span.End()
@@ -258,6 +310,30 @@ func (s *SpanReader) GetServices(ctx context.Context) ([]string, error) {
 		cfg.RolloverFrequencyAsNegativeDuration(s.serviceIndex.RolloverFrequency),
 	)
 	return s.serviceOperationStorage.getServices(ctx, jaegerIndices, s.maxDocCount)
+}
+
+// GetOperations returns all operations for a specific service traced by Jaeger
+func (s *SpanReaderV1) GetOperations(ctx context.Context, query spanstore.OperationQueryParameters) ([]spanstore.Operation, error) {
+	operations, err := s.spanReader.GetOperations(ctx, esOperationQueryParamsFromSpanStoreQueryParams(query))
+	if err != nil {
+		return nil, err
+	}
+	// TODO: https://github.com/jaegertracing/jaeger/issues/1923
+	// 	- return the operations with actual span kind that meet requirement
+	var result []spanstore.Operation
+	for _, operation := range operations {
+		result = append(result, spanstore.Operation{
+			Name: operation.Name,
+		})
+	}
+	return result, err
+}
+
+func esOperationQueryParamsFromSpanStoreQueryParams(p spanstore.OperationQueryParameters) dbmodel.OperationQueryParameters {
+	return dbmodel.OperationQueryParameters{
+		ServiceName: p.ServiceName,
+		SpanKind:    p.SpanKind,
+	}
 }
 
 // GetOperations returns all operations for a specific service traced by Jaeger
@@ -298,7 +374,85 @@ func bucketToStringArray(buckets []*elastic.AggregationBucketKeyItem) ([]string,
 	return strings, nil
 }
 
-// MultiRead returns the spans attached with the traces. The first second parameter is a map between TraceIDs and LegacyTraceIds
+// FindTraces retrieves traces that match the traceQuery
+func (s *SpanReaderV1) FindTraces(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]*model.Trace, error) {
+	ctx, span := s.spanReader.tracer.Start(ctx, "FindTraces")
+	defer span.End()
+
+	uniqueTraceIDs, err := s.FindTraceIDs(ctx, traceQuery)
+	if err != nil {
+		return nil, es.DetailedError(err)
+	}
+	return s.multiRead(ctx, uniqueTraceIDs, traceQuery.StartTimeMin, traceQuery.StartTimeMax)
+}
+
+// FindTraceIDs retrieves traces IDs that match the traceQuery
+func (s *SpanReaderV1) FindTraceIDs(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]model.TraceID, error) {
+	ctx, span := s.spanReader.tracer.Start(ctx, "FindTraceIDs")
+	defer span.End()
+
+	if err := validateQuery(traceQuery); err != nil {
+		return nil, err
+	}
+	if traceQuery.NumTraces == 0 {
+		traceQuery.NumTraces = defaultNumTraces
+	}
+
+	esTraceIDs, err := s.spanReader.FindTraceIDs(ctx, esTraceQueryParamsFromSpanStoreTraceQueryParams(traceQuery))
+	if err != nil {
+		return nil, err
+	}
+
+	return convertTraceIDsStringsToModels(esTraceIDs)
+}
+
+func esTraceQueryParamsFromSpanStoreTraceQueryParams(p *spanstore.TraceQueryParameters) *dbmodel.TraceQueryParameters {
+	return &dbmodel.TraceQueryParameters{
+		ServiceName:   p.ServiceName,
+		OperationName: p.OperationName,
+		Tags:          p.Tags,
+		StartTimeMin:  p.StartTimeMin,
+		StartTimeMax:  p.StartTimeMax,
+		DurationMin:   p.DurationMin,
+		DurationMax:   p.DurationMax,
+		NumTraces:     p.NumTraces,
+	}
+}
+
+func (s *SpanReaderV1) multiRead(ctx context.Context, traceIDs []model.TraceID, startTime, endTime time.Time) ([]*model.Trace, error) {
+	ctx, childSpan := s.tracer.Start(ctx, "multiRead")
+	defer childSpan.End()
+	traceIds := make(map[dbmodel.TraceID]string)
+	for _, traceID := range traceIDs {
+		traceIds[dbmodel.TraceID(traceID.String())] = getLegacyTraceId(traceID)
+	}
+	tracesMap, err := s.spanReader.MultiRead(ctx, traceIds, startTime, endTime)
+	if err != nil {
+		return []*model.Trace{}, err
+	}
+	var traces []*model.Trace
+	for _, jsonSpans := range tracesMap {
+		spans, err := s.collectSpans(jsonSpans)
+		if err != nil {
+			err = es.DetailedError(err)
+			logErrorToSpan(childSpan, err)
+			return nil, err
+		}
+		traces = append(traces, &model.Trace{Spans: spans})
+	}
+	return traces, nil
+}
+
+func getLegacyTraceId(traceID model.TraceID) string {
+	// https://github.com/jaegertracing/jaeger/pull/1956 added leading zeros to IDs
+	// So we need to also read IDs without leading zeros for compatibility with previously saved data.
+	// TODO remove in newer versions, added in Jaeger 1.16
+	if traceID.High == 0 {
+		return strconv.FormatUint(traceID.Low, 16)
+	}
+	return fmt.Sprintf("%x%016x", traceID.High, traceID.Low)
+}
+
 func (s *SpanReader) MultiRead(ctx context.Context, traceIds map[dbmodel.TraceID]string, startTime, endTime time.Time) (map[dbmodel.TraceID][]*dbmodel.Span, error) {
 	ctx, childSpan := s.tracer.Start(ctx, "MultiRead")
 	defer childSpan.End()
@@ -361,7 +515,7 @@ func (s *SpanReader) MultiRead(ctx context.Context, traceIds map[dbmodel.TraceID
 		results, err := s.client().MultiSearch().Add(searchRequests...).Index(indices...).Do(ctx)
 		if err != nil {
 			err = es.DetailedError(err)
-			LogErrorToSpan(childSpan, err)
+			logErrorToSpan(childSpan, err)
 			return nil, err
 		}
 
@@ -376,7 +530,7 @@ func (s *SpanReader) MultiRead(ctx context.Context, traceIds map[dbmodel.TraceID
 			spans, err := s.collectJsonSpans(result.Hits.Hits)
 			if err != nil {
 				err = es.DetailedError(err)
-				LogErrorToSpan(childSpan, err)
+				logErrorToSpan(childSpan, err)
 				return nil, err
 			}
 			lastSpan := spans[len(spans)-1]
@@ -405,6 +559,47 @@ func buildTraceByIDQuery(traceID dbmodel.TraceID, legacyTraceID string) elastic.
 	return elastic.NewBoolQuery().Should(
 		elastic.NewTermQuery(traceIDField, traceIDStr).Boost(2),
 		elastic.NewTermQuery(traceIDField, legacyTraceID))
+}
+
+func convertTraceIDsStringsToModels(traceIDs []string) ([]model.TraceID, error) {
+	traceIDsMap := map[model.TraceID]bool{}
+	// https://github.com/jaegertracing/jaeger/pull/1956 added leading zeros to IDs
+	// So we need to also read IDs without leading zeros for compatibility with previously saved data.
+	// That means the input to this function may contain logically identical trace IDs but formatted
+	// with or without padding, and we need to dedupe them.
+	// TODO remove deduping in newer versions, added in Jaeger 1.16
+	traceIDsModels := make([]model.TraceID, 0, len(traceIDs))
+	for _, ID := range traceIDs {
+		traceID, err := model.TraceIDFromString(ID)
+		if err != nil {
+			return nil, fmt.Errorf("making traceID from string '%s' failed: %w", ID, err)
+		}
+		if _, ok := traceIDsMap[traceID]; !ok {
+			traceIDsMap[traceID] = true
+			traceIDsModels = append(traceIDsModels, traceID)
+		}
+	}
+
+	return traceIDsModels, nil
+}
+
+func validateQuery(p *spanstore.TraceQueryParameters) error {
+	if p == nil {
+		return ErrMalformedRequestObject
+	}
+	if p.ServiceName == "" && len(p.Tags) > 0 {
+		return ErrServiceNameNotSet
+	}
+	if p.StartTimeMin.IsZero() || p.StartTimeMax.IsZero() {
+		return ErrStartAndEndTimeNotSet
+	}
+	if p.StartTimeMax.Before(p.StartTimeMin) {
+		return ErrStartTimeMinGreaterThanMax
+	}
+	if p.DurationMin != 0 && p.DurationMax != 0 && p.DurationMin > p.DurationMax {
+		return ErrDurationMinGreaterThanMax
+	}
+	return nil
 }
 
 func (s *SpanReader) FindTraceIDs(
@@ -601,7 +796,7 @@ func (*SpanReader) buildObjectQuery(field string, k string, v string) elastic.Qu
 	return elastic.NewBoolQuery().Must(keyQuery)
 }
 
-func LogErrorToSpan(span trace.Span, err error) {
+func logErrorToSpan(span trace.Span, err error) {
 	span.RecordError(err)
 	span.SetStatus(codes.Error, err.Error())
 }
